[IABT-1490] Fix amount text a11y in `ListItemTransaction` (#91)
## Short description
This PR fixes an accessibility issue with the `ListItemTransaction`
where, on iOS, VoiceOver reads time and amount numbers as one number
(eg: `23:15 45,67€` reads "23" and "1.545,67€").

The fix consists in setting to `false` the `accessible` prop in the
parent component. Now the `accessible` prop is passed from the parent
component, default is `false`.

## List of changes proposed in this pull request
-
[src/components/listitems/ListItemTransaction.tsx](https://github.com/pagopa/io-app-design-system/pull/91/files#diff-e161c812707e7a718a116a2c9a28e50d9d46a9956ffa18d7950aafe270570ed9):
added the `accessible` prop in the parent component.

## How to test
Use VoiceOver on iOS to read an `ListItemTransaction` component.
This issue should be tests on physical devices since it cannot be
reproduced on simulators.