[IOAPPX-303] Add the first iteration of `SearchInput`, a specific `TextInput` for search (#273)
## Short description
This PR adds the first iteration of `SearchInput`, a specific
`TextInput` designed for search.

## List of changes proposed in this pull request
- Add the new `SearchInput` component with both legacy and new visual
styles
- Add the relative screen to the _example app_

> [!important]
> If you want to clear the text, you can press the icon located to the
right. This behavior was developed using a `Pressable` custom componen
and it's available in both iOS and Android. To ensure predictable button
behavior when the virtual keyboard is visible, you need to add the prop
and value `keyboardShouldPersistTaps="always"` to the `ScrollView` that
contains the component.

### Preview

https://github.com/pagopa/io-app-design-system/assets/1255491/8e21e29c-7135-4ce8-af16-65f4419b176b


## How to test
1. Launch the example app
2. Go to the **Search Input** page

---------

Co-authored-by: Alessandro Dell'Oste <alessandro.delloste1@gmail.com>