[IOAPPX-421] Make `ListItemTransaction` independent of transaction status (#353)
## Short description
This PR removes the specific `status` prop related to
`ListItemTransaction`, as this kind of logic should be applied in the
main `io-app` repository. Making it agnostic has also improved the code
readability.

## List of changes proposed in this pull request
- Remove `status` and relative types from `ListItemTransaction`
- Add these new props to the `transaction` configuration object:
- `badge`: set the text and color variant of the badge (if set, `amount`
will be hidden)
- `refund` if you need to signal that the amount is being added and not
subtracted
- Rename the `hasChevronRight` to `showChevron`. If `true`, the text
will be rendered with the primary interactive colour

## How to test
Go to the **List Items** screen in the example app. Nothing should be
changed visually.

---------

Co-authored-by: Alessandro <alessandro.izzo@pagopa.it>