feat: support for `onBlur` and `onContinue` validation modes in `TextInputValidation` component (#405)
## Short description
This pull request introduces a new validation mode for the
`TextInputValidation` component, enabling validation based on a trigger
event

## List of changes proposed in this pull request
- Added a new `validationMode` prop to determine whether validation
occurs on blur or when an external button is pressed. Refactored the
component to use `forwardRef` and `useImperativeHandle` to expose the
`validateInput` function to parent components
- Added a new `TextInputValidationOnContinue` component that uses the
new `validationMode` prop to validate input when a "Continue" button is
pressed. Updated the `TextInputs` page to include instances of this new
component.

## How to test
- Go in `TextInputs`
- Review, test, and verify the functionality of the new
`TextInputValidation` implementation.

## Preview

https://github.com/user-attachments/assets/d905c8c6-16c5-415a-9c22-ff11e8cbbf90