feat: [SIW-2040] Collapsible list items (#407)
## Short description
This PR combines the `AccordionItem` and the `ListItemCheckbox`
components to create a collapsible component that displays a list of
items that can be selected with a checkbox or are just static.

## List of changes proposed in this pull request
- Modified `AccordionItem` to support an initial expanded state, a
secondary background color, and an optional callback on toggle
- Created `CollapsibleListItems`
- Added tests and examples

## How to test
Check the examples in Components > ListItems.
Check Components > Accordion to ensure no regressions were introduced.


https://github.com/user-attachments/assets/592ca09a-19e8-4c43-bd28-138ebb6f75d1

---------

Co-authored-by: Damiano Plebani <damiano.plebani@pagopa.it>