feat: [IOPLT-966, IOPLT-991] Add the new `IOSkeleton` component to remove the `rn-placeholder` dependency (#427)
## Short description
This PR adds the new `IOSkeleton` component to remove the obsolete and
abandoned `rn-placeholder` dependency.

### Features
Here are the features, compared to the legacy dependency:
- The new component supports only the `pulse` animation
- It's based on the `reanimated` library
- It supports the _Reduced motion_ system setting to improve
accessibility
- If it's enabled, the pulse animation is disabled and the blocks are
rendered with a static gray shade
- It includes support for dark mode

## List of changes proposed in this pull request
- Add and export the new `IOSkeleton` component
- Replace the legacy `Placeholder.Box` components with the new
`IOSkeleton` component
- Remove `rn-placeholder` dependency from both the library and the
example app

### Preview
In the example app you can find a screen with a series of repeated
skeletons for testing purposes:


https://github.com/user-attachments/assets/933c42f6-a8be-4f48-a142-63475df02802



## How to test
Run the example app with different devices to see if there are any
performance hiccups.