fix: [IOPLT-1049] Refactor `Avatar` and `AvatarSearch` to restore the correct behavior (#447)
## Short description
This PR refactors `Avatar` and `AvatarSearch` to correctly handle all
possible scenarios. In particular, `AvatarSearch` caused a crash in the
example app due to missing `width` and `height` parameters in the
`source` prop. `AvatarSearch` also incorrectly handled the case of
multiple elements in the array, where the first is a broken asset, but
the second is not.

## List of changes proposed in this pull request
- Update `Avatar` and `AvatarSearch` components
- Update `Logos` page with a simplified mock object
- Update the inner logic of `Avatar` because the array check is already
included in the `addCacheTimestampToUri` function

## How to test
Check `Avatar` and `AvatarSearch` in the `Logos` page

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>