chore: [IOPLT-1083] Add the new `IOButton` component that manages all variants at once (#460)
## Short description
This PR adds the new `IOButton` component that manages all variants at
once. You no longer need to add separate `ButtonSolid`, `ButtonOutline`
and `ButtonLink` components.

## List of changes proposed in this pull request
- Add a new `IOButton` that manages all the variants at once: `solid`,
`outline` and `link`.
- **[NEW!]** Add a new `loading` state to the `outline` variant
- **[NEW!]** Add a new `danger` color to the `outline` and `link`
variants
- Some props are exclusive to the some variants:
  - `textAlign` and `numberOfLines` are exclusive to the `link` variant
- `loading` and `fullWidth` are exclusive to the block variants (`solid`
and `outline`)
- Add deprecation note to `ButtonSolid`, `ButtonOutline` and
`ButtonLink`
- Fix a visual regression about the missing color transition of the
`Icon` component on the `pressed` state
- Add relative tests for the new `IOButton` component, update the legacy
ones
- Make the logic simpler on the `Buttons` page of the example app
- Add the new **Buttons (legacy)** page to the example app

### Preview


https://github.com/user-attachments/assets/2184ea45-56a1-49dd-b75c-a19138d57f1d


## How to test
Go to the **Buttons** page in the example app and check all the buttons
in both light/dark modes. Make sure that the behavior is virtually the
same as the legacy buttons available on the **Buttons (legacy)** page.