chore: [IOPLT-1162] Remove `storybook` (#480)
## Short description
This PR removes all traces of `storybook` from the repository. To
preserve the documentation managed within the tool, some local READMEs
have been edited or added to the repository.

## List of changes proposed in this pull request
- Remove `storybook`
- Update the main `README`
- Move the contexts from `core` to the new `context` folder (and update
all the paths accordingly)
- Move the headers from the `layout` folder to the new `headers` folder
- Move the screen templates from the `layout` folder to the new
`templates` folder

These new folders have been created to keep the logic blocks of the
documentation separate.

## How to test
Check the example app on both iOS and Android devices. Test whether the
new paths have been exported correctly by using this PR as a replacement
for the `io-app-design-system` in the `io-app`.

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>