feat: [IOBP-1987] Support badge as top element in `ListItemInfo` (#495)
## Short description
This pull request introduces support for displaying a badge as a "top
element" in the `ListItemInfo` component. The changes include updates to
both the component's type definitions and its rendering logic, as well
as new usage examples in the documentation.

## List of changes proposed in this pull request

* Added a new `topElement` prop to the `ListItemInfo` component,
allowing a badge to be rendered above the value. Updated the component's
rendering logic to display the badge when provided.
* Added new examples to the documentation demonstrating the use of
`ListItemInfo` with a badge as the top element, including different
badge variants.

## Preview
<img width="382" height="93" alt="image"
src="https://github.com/user-attachments/assets/8c1e5ecb-4d46-447a-b807-243d041e3706"
/>