chore(IT Wallet): [SIW-3134] Add badge with chevron in module credential (#505)
## Short description
This PR introduces a new optional prop to `ModuleCredential` that allows
displaying the chevron icon even when a badge is present.

## List of changes proposed in this pull request
- Added the `showChevronWithBadge`  optional prop to `ModuleCredential`
- Updated internal layout logic to render the chevron alongside the
badge when the prop is true
- Added a case with `showChevronWithBadge` in `Module.tsx`

## How to test
In `ModuleCredential` verify that:
- The chevron is not shown by default when a badge is present.
- When `showChevronWithBadge={true}` is passed, both the badge and the
chevron are visible and properly aligned.

<img width="401" height="892" alt="Screenshot 2025-10-09 at 15 32 32"
src="https://github.com/user-attachments/assets/6fe40783-15b0-4b1c-b72b-ace4aebd3edc"
/>

---------

Co-authored-by: Damiano Plebani <damiano.plebani@pagopa.it>